greedy

+++ item.hpp
@@ -1,12 +1,14 @@
 #ifndef ITEM_HPP
 #define ITEM_HPP
 
+#include <cstdint>
+
 struct item
 {
-    item(int value, int weight) : value(value), weight(weight){};
+    item(int64_t value, int64_t weight) : value(value), weight(weight){};
 
-    int value;
-    int weight;
+    int64_t value;
+    int64_t weight;
 };
 
 #endif  // !ITEM_HPP
